+++ R/method.A.R
@@ -41,10 +41,10 @@ method.A <- function(alpha = 0.05, path.in = "~/", path.out = "~/",
     name <-  paste0(file, set)
     cat("\nData set", name, "by Method A",
         paste0("\n", paste0(rep("\u2500", 22+nchar(name)), collapse="")), "\n")
-    print(stats::anova(modA), digits=7) # otherwise summary of lmerTest is used
+    print(stats::anova(modA), digits=7) # otherwise summary of lmerTest is used, digits=7 ignored
     cat("\ntreatment T \u2013 R:\n")
     print(signif(summary(modA)$coefficients["treatmentT", ]), 6)
-    cat(summary(modA)$df[2], "Residual Degrees of Freedom\n\n")
+    cat(summary(modA)$df[2], "Degrees of Freedom\n\n")
   }
   PE  <- exp(coef(modA)[["treatmentT"]])
   CI  <- as.numeric(exp(confint(modA, "treatmentT", level=1-2*alpha)))