+++ NEWS.md
@@ -1,6 +1,10 @@
 # replicateBE 1.0.12.9000
 
-Published on Github 2019-12-25 (passed checks with R-release and R-devel on win-builder).
+Published on Github 2019-12-26 (passed 2019-12-25 on win-builder checks with R-release and R-devel).
+
+## Issues
+
+  * `digits = 7` is ignored in `print(anova(foo, ...))`. Why?
 
 ## Bug fixes
 
@@ -8,6 +12,7 @@ Published on Github 2019-12-25 (passed checks with R-release and R-devel on win-
 
 ## Minor changes
 
+  * Unified ANOVA-table and degrees of freedom if `verbose = TRUE`. Added response to verbose-output of `method.B` (automatically in `method.A`).
   * Handles a case where a user exports one of the internal dataset as a CSV *with* row.names and quoted variables (*i.e.*, instead of `write.csv(rds01, 'rds01.csv', quote = FALSE, row.names = FALSE`) naïvely uses just `write.csv(rds01, 'rds01.csv')`).  
   Importing such a CSV-file was not possible so far. Resolved [issue 3](https://github.com/Helmut01/replicateBE/issues/3) raised by MT.
   * Adapted tests for `method.B()`: `rds29` and `rds30` instead of `rds18`. Satterthwaite DF `(..., option = 1)` and Kenward-Roger `(..., option = 3)`.